remove unneeded copies from ds6-safety.cpp fixes wrappers/python/c_files.cpp

+++ src/ds/ds6/ds6-private.h
@@ -15,8 +15,8 @@ namespace librealsense
         // Safety depth XU identifiers
         namespace xu_id
         {
-            const uint16_t SAFETY_CAMERA_OPER_MODE = 0x1;
-            const uint16_t SAFETY_PRESET_ACTIVE_INDEX = 0x2;
+            const uint8_t SAFETY_CAMERA_OPER_MODE = 0x1;
+            const uint8_t SAFETY_PRESET_ACTIVE_INDEX = 0x2;
         }
 
         // ds6 Devices supported by the current version
